requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.4
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.3.1
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
